+++ src/realm/sync/client.cpp
@@ -638,8 +638,8 @@ void ClientImpl::remove_connection(ClientImpl::Connection& conn) noexcept
 // ################ SessionImpl ################
 
 
-inline void SessionImpl::on_connection_state_changed(ConnectionState state,
-                                                     const util::Optional<SessionErrorInfo>& error_info)
+void SessionImpl::on_connection_state_changed(ConnectionState state,
+                                              const util::Optional<SessionErrorInfo>& error_info)
 {
     m_wrapper.on_connection_state_changed(state, error_info); // Throws
 }
@@ -1403,13 +1403,6 @@ void SessionWrapper::on_resumed()
 void SessionWrapper::on_connection_state_changed(ConnectionState state,
                                                  const util::Optional<SessionErrorInfo>& error_info)
 {
-    if (state == ConnectionState::connected && m_sess) {
-        ClientImpl::Connection& conn = m_sess->get_connection();
-        if (conn.is_flx_sync_connection()) {
-            get_flx_subscription_store();
-        }
-    }
-
     if (m_connection_state_change_listener) {
         if (!m_suspended)
             m_connection_state_change_listener(state, error_info); // Throws